refactor(game-page): drop per-source split from hero, surface on stats page instead
"App XXX · newest report: N days ago · 0 Pulse / 4 ProtonDB" was
overloaded and the trailing bit ate valuable horizontal space on
mobile. Hero now shows just app id + freshness.

The N Pulse / M ProtonDB split moved into the stats page header
sub-line where it belongs (it's a stats page thing, not hero
information). renderHeader now accepts { pulseCount, protonDbCount }
and appends when either is non-zero.
