+++ game-stats.html
@@ -335,7 +335,7 @@ h1 {
 <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js@2/dist/umd/supabase.min.js"></script>
 <script src="js/lib/supabase-client.js?v=79bf62c6"></script>
 <script src="js/lib/analytics.js?v=e6dfbad8"></script>
-<script type="module" src="js/game-stats/main.js?v=bd125960"></script>
+<script type="module" src="js/game-stats/main.js?v=ab9c5267"></script>
 
 </body>
 </html>
